Cleanup styles and redux documents
