--- client/mobile/.expo/packager-info.json
@@ -1,8 +1,0 @@
-{
-  "expoServerPort": null,
-  "expoServerNgrokUrl": null,
-  "packagerNgrokUrl": null,
-  "ngrokPid": null,
-  "packagerPort": null,
-  "packagerPid": null
-}
