--- client/mobile/.expo/settings.json
@@ -1,7 +1,0 @@
-{
-  "hostType": "tunnel",
-  "lanType": "ip",
-  "dev": true,
-  "minify": false,
-  "urlRandomness": "kh-z9x"
-}
