+++ client/mobile/components/home/home.js
@@ -1,14 +1,6 @@
 import React from 'react';
-import { View, Button, Text, StyleSheet } from 'react-native';
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    backgroundColor: '#fff',
-    alignItems: 'center',
-    justifyContent: 'center'
-  },
-});
+import { View, Button, Text } from 'react-native';
+import styles from '../../styles';
 
 const Home = ({ createDoc }) => (
   <View style={styles.container}>
