+++ client/mobile/components/splash/welcome.js
@@ -1,14 +1,6 @@
 import React from 'react';
-import { StyleSheet, Text, View, Button } from 'react-native';
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    backgroundColor: '#fff',
-    alignItems: 'center',
-    justifyContent: 'center'
-  }
-});
+import { Text, View, Button } from 'react-native';
+import styles from '../../styles';
 
 const Welcome = ({ navigation }) => (
   <View style={styles.container}>
