+++ client/shared/reducers/root_reducer.js
@@ -1,7 +1,7 @@
 import { combineReducers } from 'redux';
 import session from './session_reducer';
 import errors from './errors_reducer';
-import documents from './document_reducer';
+import documents from './documents_reducer';
 import ui from './ui_reducer';
 import users from './users_reducer';
